examples/06_netauto_ecosystem/06_nornir_project/requirements.txt

nornir>=3.4.1
nornir-scrapli>=2023.7.30
scrapli>=2024.0.0
PyYAML>=6.0.1